+++ Editor/Playa/Renderer/MethodBindFakeRenderer/MethodBindRendererUIToolkit.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using SaintsField.Editor.Core;
 using SaintsField.Editor.UIToolkitElements;
 using SaintsField.Editor.Utils;
+using SaintsField.Playa;
 using UnityEditor;
 using UnityEngine.Events;
 using UnityEngine.UIElements;
@@ -85,9 +86,19 @@ namespace SaintsField.Editor.Playa.Renderer.MethodBindFakeRenderer
                 }
             }
 
-            string targetEventName = _methodBindAttribute.EventTarget == null
-                ? "On Click"
-                : $"{_methodBindAttribute.EventTarget.Split('.').Last()}{eventSuffix}";
+            string targetEventName;
+            if (_methodBindAttribute.MethodBind == MethodBind.ComponentTypeAndName)
+            {
+                targetEventName = _methodBindAttribute.ComponentEventName;
+            }
+            else if (_methodBindAttribute.EventTarget == null)
+            {
+                targetEventName = "On Click";
+            }
+            else
+            {
+                targetEventName = $"{_methodBindAttribute.EventTarget.Split('.').Last()}{eventSuffix}";
+            }
 
             _onEventWithErrorElement = new OnEventWithErrorElement(ObjectNames.NicifyVariableName(FieldWithInfo.MethodInfo.Name), targetEventName, $"{FieldWithInfo.MemberId}[{targetEventName}]");
             _onEventWithErrorElement.RegisterCallback<AttachToPanelEvent>(_ =>