+++ Editor/SaintsEditor.cs
@@ -1273,16 +1273,10 @@ namespace SaintsField.Editor
                         }
                     }
                         break;
-                    case OnButtonClickAttribute onButtonClickAttribute:
+                    case IPlayaMethodBindAttribute methodBindAttribute:
                     {
                         // yield return new SaintsFieldWithRenderer(onButtonClickAttribute, new MethodBindRenderer(onButtonClickAttribute, serializedObject, fieldWithInfo));
-                        preRenderers.Add(new SaintsFieldWithRenderer(onButtonClickAttribute, new MethodBindRenderer(onButtonClickAttribute, serializedObject, fieldWithInfo)));
-                    }
-                        break;
-                    case OnEventAttribute onEventAttribute:
-                    {
-                        // yield return new SaintsFieldWithRenderer(onEventAttribute, new MethodBindRenderer(onEventAttribute, serializedObject, fieldWithInfo));
-                        preRenderers.Add(new SaintsFieldWithRenderer(onEventAttribute, new MethodBindRenderer(onEventAttribute, serializedObject, fieldWithInfo)));
+                        preRenderers.Add(new SaintsFieldWithRenderer(methodBindAttribute as IPlayaAttribute, new MethodBindRenderer(methodBindAttribute, serializedObject, fieldWithInfo)));
                     }
                         break;
                     case InfoBoxAttribute playaInfoBoxAttribute: