+++ Runtime/Playa/MethodBind.cs
@@ -3,8 +3,6 @@ namespace SaintsField.Playa
     public enum MethodBind
     {
         UnityEvent,
-        ButtonOnClick,
-        ToggleOnValueChanged,
-        TMPDropdownOnValueChanged,
+        ComponentTypeAndName,
     }
 }