+++ Runtime/Playa/OnButtonClickAttribute.cs
@@ -8,9 +8,11 @@ namespace SaintsField.Playa
     [AttributeUsage(AttributeTargets.Method)]
     public class OnButtonClickAttribute: Attribute, IPlayaAttribute, IPlayaMethodAttribute, IPlayaMethodBindAttribute, IPlayaAutoRunnerFix
     {
-        public MethodBind MethodBind => MethodBind.ButtonOnClick;
+        public MethodBind MethodBind => MethodBind.ComponentTypeAndName;
 
         public string EventTarget { get; }
+        public Type ComponentTypeOrNull => typeof(UnityEngine.UI.Button);
+        public string ComponentEventName => nameof(UnityEngine.UI.Button.onClick);
         public object Value { get; }
         public bool IsCallback { get; }