+++ Runtime/Playa/OnEventAttribute.cs
@@ -11,6 +11,8 @@ namespace SaintsField.Playa
         public MethodBind MethodBind => MethodBind.UnityEvent;
 
         public string EventTarget { get; }
+        public Type ComponentTypeOrNull => null;
+        public string ComponentEventName => null;
         public object Value { get; }
         public bool IsCallback { get; }